Fix home trail gets into LD schema
